Fixed an issue using 'test_split_Depression_AVEC2017.csv' in test mode as there are no labels for PHQ8_Binary. This is not the case with 'full_test_split.csv', which will work the same as a standard validation set in test mode.
If we are working with a hidden test set, we cannot use prediction_metric=0 (pick best model) as this is done by calculating the accuracy and F-Score. Instead, when working in test mode with the hidden test set ('test_split_Depression_AVEC2017.csv') we either average or majority vote the outputs of the network and present the predictions alone as output.